Modularize app.
