+++ app.js
@@ -1,10 +1,14 @@
 'use strict';
 
-const http         = require('http');
 const path         = require('path');
 const express      = require('express');
 const uuidv4       = require('uuid/v4');
 const semver       = require('semver');
+
+const helpers      = require('./lib/helpers.js');
+const config       = helpers.getConfig();
+
+const app          = express();
 
 // constants
 const ENV          = process.env;
@@ -29,15 +33,10 @@ const staticify    = require('staticify')(PUBLIC_DIR, {
     sendOptions: STATIC_OPTS
 });
 
-const CSP          = require('./config/helmet-csp.js');
-const helpers      = require('./lib/helpers.js');
-const routes       = require('./routes');
-
-const config       = helpers.getConfig();
-const app          = express();
+const CSP = require('./config/helmet-csp.js');
+const routes = require('./routes/');
 
 // all environments
-app.set('port', ENV.PORT || config.port || 3000);
 app.set('views', path.join(__dirname, '/views/'));
 app.set('view engine', 'pug');
 app.set('etag', false);
@@ -162,27 +161,21 @@ app.locals.getVersionedPath = staticify.getVersionedPath;
 app.locals.semver = semver;
 
 // routes
-app.get('/', routes.renderIndex);
-app.get('/about/', routes.renderAbout);
-app.get('/alpha/', routes.redirectToRoot);
-app.get('/beta/', routes.redirectToRoot);
-app.get('/bootlint/', routes.renderBootlint);
-app.get('/bootswatch/', routes.renderBootswatch);
-app.get('/bootswatch4/', routes.renderBootswatch4);
-app.get('/fontawesome/', routes.renderFontawesome);
-app.get('/integrations/', routes.renderIntegrations);
-app.get('/legacy/', routes.legacy);
-app.get('/legacy/bootstrap/', routes.renderLegacyBootstrap);
-app.get('/legacy/bootswatch/', routes.renderLegacyBootswatch);
-app.get('/legacy/fontawesome/', routes.renderLegacyFontawesome);
-app.get('/privacy-policy/', routes.renderPrivacyPolicy);
-app.get('/showcase/', routes.renderShowcase);
-
-app.get('/data/bootstrapcdn.json', (req, res) => {
-    const data = helpers.generateDataJson();
-
-    res.send(data);
-});
+app.use('/', routes.indexRoute);
+app.use('/about/', routes.aboutRoute);
+app.use('/alpha/?|/beta/?', routes.redirectToRoot);
+app.use('/bootlint/', routes.bootlintRoute);
+app.use('/bootswatch/', routes.bootswatchRoute);
+app.use('/bootswatch4/', routes.bootswatch4Route);
+app.use('/data/bootstrapcdn.json', routes.dataRoute);
+app.use('/fontawesome/', routes.fontawesomeRoute);
+app.use('/integrations/', routes.integrationsRoute);
+app.use('/legacy/', routes.legacyRoute);
+app.use('/legacy/bootstrap/', routes.legacyBootstrapRoute);
+app.use('/legacy/bootswatch/', routes.legacyBootswatchRoute);
+app.use('/legacy/fontawesome/', routes.legacyFontawesomeRoute);
+app.use('/privacy-policy/', routes.privacyPolicyRoute);
+app.use('/showcase/', routes.showcaseRoute);
 
 const map = sitemap({
     url: 'www.bootstrapcdn.com',
@@ -220,9 +213,8 @@ if (ENV.ENABLE_CRAWLING) {
 
 app.get('/robots.txt', (req, res) => map.TXTtoWeb(res));
 
-app.get('*', routes.render404);
+app.use('*', routes.notFoundRoute);
 
-// start
-http.createServer(app).listen(app.get('port'), () => console.log(`Express server listening on port ${app.get('port')}`));
+module.exports = app;
 
 // vim: ft=javascript sw=4 sts=4 et:
