+++ routes/legacy.js
@@ -1,0 +1,15 @@
+'use strict';
+
+const express = require('express');
+const appendLocals = require('./appendLocals.js');
+
+const router = express.Router();
+
+router.get('/', (req, res) => {
+    res = appendLocals(req, res);
+    res.redirect(301, '/legacy/bootstrap/');
+});
+
+module.exports = router;
+
+// vim: ft=javascript sw=4 sts=4 et:
