+++ routes/showcase.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const express = require('express');
+const appendLocals = require('./appendLocals.js');
+
+const router = express.Router();
+
+router.get('/', (req, res) => {
+    res = appendLocals(req, res);
+    res.render('showcase.pug', {
+        title: 'Showcase',
+        description: 'Websites and apps that use BootstrapCDN'
+    });
+});
+
+module.exports = router;
+
+// vim: ft=javascript sw=4 sts=4 et:
