+++ test/test_helpers.js
@@ -82,7 +82,7 @@ function runApp(cfg, endpoint) {
     process.env.PORT = port;
 
     // load app
-    require('../app.js');
+    require('../bin/www.js');
 
     return `http://localhost:${port}${endp}`;
 }
